# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "own.audio server API",
    "description": "The contract every own.audio client speaks. Additive within v1: see docs/API_COMPATIBILITY.md.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "1.4"
  },
  "paths": {
...
    "/api/v1/library/folders": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "The family's configured library folders and how their last scan went",
        "operationId": "library_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoldersResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a family admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/library/folders/scan": {
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Start a scan of the library folders now",
        "operationId": "library_scan",
        "responses": {
          "202": {
            "description": "Scan requested"
          },
          "403": {
            "description": "Not a family admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Every error answer: a status code and this body.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in English; not meant to be matched on."
          }
        }
      },
...
      "FolderStatus": {
        "type": "object",
        "required": [
          "id",
          "path",
          "kind",
          "visibility",
          "files_seen",
          "files_added",
          "files_missing"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          },
          "scan_started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "scan_finished_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "scan_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "files_seen": {
            "type": "integer",
            "format": "int32"
          },
          "files_added": {
            "type": "integer",
            "format": "int32"
          },
          "files_missing": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FoldersResponse": {
        "type": "object",
        "required": [
          "scanning",
          "files_scanned",
          "folders"
        ],
        "properties": {
          "scanning": {
            "type": "boolean"
          },
          "files_scanned": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FolderStatus"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The access token from `POST /auth/login` or `POST /auth/refresh`."
      }
    }
  }
}